Moved embedding_biology.py to the examples folder
