+++ examples/weaviate/vector_search.py
@@ -1,7 +1,8 @@
-from neo4j_genai.retrievers.external.weaviate import WeaviateNeo4jRetriever
 from neo4j import GraphDatabase
+from neo4j_genai.retrievers.external.weaviate import WeaviateNeo4jRetriever
 from weaviate.connect.helpers import connect_to_local
-from embedding_biology import EMBEDDING_BIOLOGY
+
+from examples.embedding_biology import EMBEDDING_BIOLOGY
 
 NEO4J_URL = "neo4j://localhost:7687"
 NEO4J_AUTH = ("neo4j", "password")
